+++ exercise1/part2/solution/Kelvin p2.py
