+++ exercise1/part3/solution/Kelvin p3.py
