+++ exercise1/part4/solution/Kelvin p4.py
