+++ exercise2/Vending.py
@@ -1,0 +1,33 @@
+def ShowMenu():
+    print('Vending Machine')
+    print('+-------------+-------------+-------------+')
+    print('|    Water    |     Coke    |  Diet Coke  |')
+    print('|    $0.75    |    $1.20    |    $1.20    |')
+    print('|     A1      |     A2      |     A3      |')
+    print('+-------------+-------------+-------------+')
+    print('|  Iced Tea   |  Chocolate  |    Candy    |')
+    print('|    $1.00    |    $1.50    |    $0.95    |')
+    print('|     B1      |     B2      |     B3      |')
+    print('+-------------+-------------+-------------+')
+    print('|    Chips    |     Gum     |Turk. Delight|')
+    print('|    $1.10    |    $0.50    |    $1.20    |')
+    print('|     C1      |     C2      |     C3      |')
+    print('+-------------+-------------+-------------+')
+    print('|        +----------------------+         |')
+    print('|        |                      |         |')
+    print('|        +----------------------+         |')
+    print('+-----------------------------------------+')
+
+def MakeSelection(sum):
+    selection = input('Please select a product by typing its code or enter q to quit')
+    if selection =
+
+
+
+
+def ReturnChange(amount):
+
+
+
+
+def DisplayErrorMessage()
